[NPU L0] update layernorm & code refactor (#12287)
* update layernorm & code refactor

* fix style

* add common utils

* change to Pool()

* remove print
